+++ packages/shared/src/types/renderers.types.ts
@@ -1,0 +1,3 @@
+export enum RendererType {
+  Markdown = 'MARKDOWN',
+}
